добавлена фабрика контекстов БД

+++ BoardGameBot/BoardGameBot.Database.PostgreSQL/Migrations/BoardGameContextModelSnapshot.cs
@@ -24,11 +24,11 @@ namespace BoardGameBot.Database.PostgreSQL.Migrations
 
             modelBuilder.Entity("BoardGameBot.Database.PostgreSQL.Models.Game", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("integer");
+                        .HasColumnType("bigint");
 
-                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
 
                     b.Property<int>("Complexity")
                         .HasColumnType("integer");
@@ -70,17 +70,17 @@ namespace BoardGameBot.Database.PostgreSQL.Migrations
 
             modelBuilder.Entity("BoardGameBot.Database.PostgreSQL.Models.GameOwner", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("integer");
+                        .HasColumnType("bigint");
 
-                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
 
-                    b.Property<int>("GroupAdminId")
-                        .HasColumnType("integer");
+                    b.Property<long>("GroupAdminId")
+                        .HasColumnType("bigint");
 
-                    b.Property<int>("GroupMemberId")
-                        .HasColumnType("integer");
+                    b.Property<long>("GroupMemberId")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("Name")
                         .IsRequired()
@@ -104,11 +104,11 @@ namespace BoardGameBot.Database.PostgreSQL.Migrations
 
             modelBuilder.Entity("BoardGameBot.Database.PostgreSQL.Models.Group", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("integer");
+                        .HasColumnType("bigint");
 
-                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
 
                     b.Property<string>("Description")
                         .IsRequired()
@@ -128,18 +128,18 @@ namespace BoardGameBot.Database.PostgreSQL.Migrations
 
             modelBuilder.Entity("BoardGameBot.Database.PostgreSQL.Models.Poll", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("integer");
+                        .HasColumnType("bigint");
 
-                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
 
                     b.Property<int>("DayInterval")
                         .HasMaxLength(1)
                         .HasColumnType("integer");
 
-                    b.Property<int>("GroupId")
-                        .HasColumnType("integer");
+                    b.Property<long>("GroupId")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("Name")
                         .IsRequired()
@@ -160,11 +160,11 @@ namespace BoardGameBot.Database.PostgreSQL.Migrations
 
             modelBuilder.Entity("GameGameOwner", b =>
                 {
-                    b.Property<int>("GameOwnersId")
-                        .HasColumnType("integer");
+                    b.Property<long>("GameOwnersId")
+                        .HasColumnType("bigint");
 
-                    b.Property<int>("GamesId")
-                        .HasColumnType("integer");
+                    b.Property<long>("GamesId")
+                        .HasColumnType("bigint");
 
                     b.HasKey("GameOwnersId", "GamesId");
